+++ Api/Api/Dtos/Championship.cs
@@ -4,6 +4,7 @@ namespace Api.Dtos;
 
 public class GetPossibleChampionshipsDto
 {
+    public int Id { get; set; }
     public string Name { get; set; }
     public string LogoUrl { get; set; }
     public string Region { get; set; }