+++ Client/src/lib/components/Header/Header.svelte
@@ -8,6 +8,9 @@
   export let headerButtonFadingOut = false;
   export let headerButtonVisible = true;
   export let hiddenMenu = false;
+  export let toggleMenuOpen = () => {
+    null;
+  };
 
   let showLoginModal = false;
   let showRegisterModal = false;
@@ -94,8 +97,8 @@
   }
 </script>
 
-<div
-  class="fixed left-0 top-0 z-40 flex h-16 w-full items-center justify-between bg-dark-grey px-4 shadow-md shadow-black md:px-10 "
+<header
+  class="fixed left-0 top-0 z-40 flex h-16 w-full items-center justify-between bg-dark-grey px-4 shadow-md shadow-black"
 >
   <img src={Logo} alt="PreCog logo" />
   <div class="flex gap-4">
@@ -123,7 +126,7 @@
     />
   </button>
   </div>
-</div>
+</header>
 
 <Modal bind:showModal="{showLoginModal}" bind:closeModal={closeLoginModal}>
   <div slot="body" class="grid grid-cols-1 content-center">