+++ Client/src/lib/components/Menu/Menu.story.svelte
@@ -5,12 +5,12 @@
   export let menuOpen: boolean
   export let menuClosing: boolean
 
-  export let Hst: Hst;
+  let Hst: Hst;
 </script>
 
-<Hst.Story title="Menu">
+<svelte:component this={Hst.Story} title="Menu">
   <svelte:component this={Hst.Variant}>
     <Menu {menuOpen} {menuClosing}/>
     <svelte:fragment slot="controls"></svelte:fragment>
   </svelte:component>
-</Hst.Story>
+</svelte:component>