+++ Api/Api/CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Api project
 - Build job on github actions
+- [API-42] Predict route